requirements.txt
aiohttp==3.6.3
arkecosystem-crypto==1.0.0
arrow==0.17.0
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==20.3.0
backpack==0.1
base58==2.0.1
binary-helpers==0.0.4
blinker==1.4
cffi==1.14.4
chardet==3.0.4
cleo==0.6.8
coincurve==13.0.0
discord.py==1.5.1
Faker==0.8.18
idna==2.10
inflection==0.3.1
lazy-object-proxy==1.5.2
loguru==0.5.3
multidict==4.7.6
orator==0.9.9
pastel==0.1.1
pendulum==1.5.1
psycopg2-binary==2.8.6
py-database-url==0.0.4
pyaml==16.12.2
pycparser==2.20
Pygments==2.7.2
pylev==1.3.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
pytzdata==2020.1
PyYAML==5.3.1
simplejson==3.17.2
six==1.15.0
text-unidecode==1.2
tzlocal==1.5.1
wrapt==1.12.1
yarl==1.5.1
